feat(RF05): Add update task page
