+++ src/pages/Tasks/update.tsx
@@ -1,18 +1,21 @@
 import { useEffect, useState } from 'react';
-// import { useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import UpdateTaskForm from '../../components/modules/Task/UpdateTask/UpdateTaskForm';
 import useHttp from '../../hooks/useHttp';
 import { EmployeeEntity } from '../../types/employee';
 import { Response } from '../../types/response';
-import { UpdatedTask } from '../../types/task';
-import { RequestMethods } from '../../utils/constants';
+import { TaskDetail, UpdatedTask } from '../../types/task';
+import { APIPath, RequestMethods } from '../../utils/constants';
 
 const UpdateTaskPage: React.FC = () => {
+  const { id } = useParams();
+
   const { data: cachedEmployees, sendRequest: sendEmployeeRequest } = useHttp<
     Response<EmployeeEntity>
   >(`/employee/getAllEmployees`, RequestMethods.GET);
 
   const [employees, setEmployees] = useState<EmployeeEntity[]>([]);
+
   useEffect(() => {
     sendEmployeeRequest();
   }, []);
@@ -23,14 +26,39 @@ const UpdateTaskPage: React.FC = () => {
     }
   }, [cachedEmployees]);
 
-  // const { idTask } = useParams();
-  const { sendRequest } = useHttp<UpdatedTask>(`/tasks/update/445564de-0f44-4634-afc1-40190242792c`, RequestMethods.PUT);
+  const { data: cachedTask, sendRequest: sendGetTaskRequest } = useHttp<TaskDetail>(
+    `${APIPath.TASK_DETAIL}/${id}`,
+    RequestMethods.GET
+  );
+
+  console.log(cachedTask);
+
+  useEffect(() => {
+    if (!cachedTask) {
+      sendGetTaskRequest();
+    }
+  }, []);
+
+  const { sendRequest: sendUpdateTaskRequest } = useHttp<UpdatedTask>(
+    `${APIPath.UPDATE_TASK}/${id}`,
+    RequestMethods.PUT
+  );
 
   const handleOnSubmit = async (payload: UpdatedTask) => {
-    await sendRequest({}, { ...payload });
+    try {
+      await sendUpdateTaskRequest({}, { ...payload });
+    } catch (error) {
+      console.error(error);
+    }
   };
 
-  return <UpdateTaskForm onSubmit={handleOnSubmit} employees={employees || []} />;
+  return (
+    <UpdateTaskForm
+      data={cachedTask || ({} as TaskDetail)}
+      onSubmit={handleOnSubmit}
+      employees={employees || []}
+    />
+  );
 };
 
 export default UpdateTaskPage;
